+++ core/initializer/main.go
@@ -21,6 +21,9 @@ const (
 	failureExitCode = 1
 
 	testNameArgSeparator = ","
+
+	defaultKurtosisApiImage = "kurtosistech/kurtosis-core_api:latest"
+	defaultParallelism = 4
 )
 
 func main() {
@@ -43,6 +46,16 @@ func main() {
 		"debug",
 		fmt.Sprintf("Log level string to use for the test suite (will be passed to the test suite container as-is"),
 	)
+
+	kurtosisApiImageArg := flag.String(
+		"kurtosis-api-image",
+		defaultKurtosisApiImage,
+		"The Docker image that will be used to run the Kurtosis API container")
+
+	parallelismArg := flag.Int(
+		"parallelism",
+		defaultParallelism,
+		"Number of tests to run concurrently (NOTE: should be set no higher than the number of cores on your machine!)")
 
 	// TODO add a "list tests" flag
 	flag.Parse()
@@ -73,17 +86,16 @@ func main() {
 	testSuiteRunner := test_suite_runner.NewTestSuiteRunner(
 		dockerClient,
 		*testSuiteImageArg,
-		// TODO parameterize this
-		"kurtosistech/kurtosis-core_api",
+		*kurtosisApiImageArg,
 		*testSuiteLogLevelArg,
 		// TODO parameterize this
 		map[string]string{},
 		*kurtosisLogLevelArg)
 
+	parallelismUint := uint(*parallelismArg)
 	allTestsPassed, err := testSuiteRunner.RunTests(
 		testNamesToRun,
-		// TODO parameterize this
-		4)
+		parallelismUint)
 	if err != nil {
 		logrus.Errorf("An error occurred running the tests:")
 		fmt.Fprintln(logrus.StandardLogger().Out, err)