+++ core/CHANGELOG.md
@@ -16,6 +16,7 @@
 * Remove some unnecessary `context.Context` pointer-passing
 * Made log levels of Kurtosis & test suite independently configurable
 * Switch to using CircleCI for builds
+* Made the API image & parallelism configurable
 * Remove TODO in run.sh about parameterizing binary name
 
 # 0.9.0